DDIRatioDataMapping

R6 class for mapping x, y, GroupMapping and DDI ratio lines variables to data

See also

Super classes

tlf::XYDataMapping -> tlf::XYGDataMapping -> tlf::PKRatioDataMapping -> DDIRatioDataMapping

Public fields

minRange

Mininmum range for guest and ratio lines

deltaGuest

Value of delta in Guest et al. equation

residualsVsObserved

if true, returns a mapping for generating a residuals (predicted/observed) vs observed DDI ratio plot. If false (default) returns a mapping for predicted vs observed DDI ratio plot.

Methods

Public methods

DDIRatioDataMapping$new()

DDIRatioDataMapping$clone()

Inherited methods

tlf::PKRatioDataMapping$checkMapData()

Method new()

Create a new DDIRatioDataMapping object

Usage

DDIRatioDataMapping$new(
  deltaGuest = 1,
  minRange = c(0.01, 100),
  lines = DefaultDataMappingValues$ddiRatio,
  residualsVsObserved = FALSE,
  ...
)

Arguments

deltaGuest

Value of parameter delta in Guest et al. equation. Default value is 1.

minRange

Mininmum range for guest lines

lines

List of ratio limits to display as diagonal/horizontal lines

residualsVsObserved

if true, returns a mapping for generating a residuals (predicted/observed) vs observed DDI ratio plot. If false (default) returns a mapping for predicted vs observed DDI ratio plot.

...

parameters inherited from PKRatioDataMapping

Returns

A new DDIRatioDataMapping object

Method clone()

The objects of this class are cloneable with this method.

Usage

DDIRatioDataMapping$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.