LabelConfiguration

R6 class defining the configuration of the labels of a ggplot object

See also

Active bindings

title

Label object defining the title of the plot

subtitle

Label object defining the subtitle of the plot

xlabel

Label object defining the xlabel of the plot

ylabel

Label object defining the ylabel of the plot

caption

Label object defining the caption of the plot

y2label

Label object defining the y2label of the plot

Methods

Public methods

LabelConfiguration$new()

LabelConfiguration$updatePlot()

LabelConfiguration$clone()

Method new()

Create a new LabelConfiguration object

Usage

LabelConfiguration$new(
  title = NULL,
  subtitle = NULL,
  xlabel = NULL,
  ylabel = NULL,
  caption = NULL
)

Arguments

title

character or Label object defining title

subtitle

character or Label object defining subtitle

xlabel

character or Label object defining xlabel

ylabel

character or Label object defining ylabel

caption

character or Label object defining caption

Returns

A new LabelConfiguration object

Method updatePlot()

Update labels of a ggplot object and their properties

Usage

LabelConfiguration$updatePlot(plotObject)

Arguments

plotObject

a ggplot object

Returns

A ggplot object

Method clone()

The objects of this class are cloneable with this method.

Usage

LabelConfiguration$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.