ObservedDataMapping

R6 class for mapping x, y, of observed data for a time profile plot

See also

Super classes

tlf::XYDataMapping -> tlf::XYGDataMapping -> ObservedDataMapping

Public fields

error

mapping error bars around scatter points

mdv

mapping missing dependent variable

ymin

mapping error bars around scatter points

ymax

mapping error bars around scatter points

y2Axis

Name of y2Axis variable to map

Methods

Public methods

ObservedDataMapping$new()

ObservedDataMapping$checkMapData()

ObservedDataMapping$requireDualAxis()

ObservedDataMapping$getLeftAxis()

ObservedDataMapping$getRightAxis()

ObservedDataMapping$clone()

Method new()

Create a new ObservedDataMapping object

Usage

ObservedDataMapping$new(
  x,
  y,
  ymin = NULL,
  ymax = NULL,
  y2Axis = NULL,
  group = NULL,
  color = NULL,
  shape = NULL,
  error = NULL,
  uncertainty = NULL,
  mdv = NULL,
  data = NULL
)

Arguments

x

Name of x variable to map

y

Name of y variable to map

ymin

mapping lower end of error bars around scatter points

ymax

mapping upper end of error bars around scatter points

y2Axis

Name of y2Axis variable to map

group

R6 class Grouping object or its input

color

R6 class Grouping object or its input

shape

R6 class Grouping object or its input

error

mapping error bars around scatter points

uncertainty

mapping error bars around scatter points. Deprecated parameter replaced by error.

mdv

mapping missing dependent variable

data

data.frame to map used by .smartMapping

Returns

A new ObservedDataMapping object

Method checkMapData()

Check that data variables include map variables

Usage

ObservedDataMapping$checkMapData(data, metaData = NULL)

Arguments

data

data.frame to check

metaData

list containing information on data

Returns

A data.frame with map and defaultAes variables. Dummy variable defaultAes is necessary to allow further modification of plots.

Method requireDualAxis()

Assess if data require a dual axis plot

Usage

ObservedDataMapping$requireDualAxis(data)

Arguments

data

data.frame to check

Returns

A logical

Method getLeftAxis()

Render NA values for all right axis data

Usage

ObservedDataMapping$getLeftAxis(data)

Arguments

data

A data.frame

Returns

A data.frame to be plotted in left axis

Method getRightAxis()

Render NA values for all left axis data

Usage

ObservedDataMapping$getRightAxis(data)

Arguments

data

A data.frame

Returns

A data.frame to be plotted in right axis

Method clone()

The objects of this class are cloneable with this method.

Usage

ObservedDataMapping$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.