QQDataMapping

R6 class for mapping x, y and GroupMapping variables to data

See also

Super classes

tlf::XYDataMapping -> tlf::XYGDataMapping -> QQDataMapping

Methods

Public methods

QQDataMapping$checkMapData()

QQDataMapping$clone()

Inherited methods

tlf::XYGDataMapping$initialize()

Method checkMapData()

Check that data variables include map variables

Usage

QQDataMapping$checkMapData(data, metaData = NULL)

Arguments

data

data.frame to check

metaData

list containing information on data

Returns

A data.frame with map and defaultAes variables. Dummy variable defaultAes is necessary to allow further modification of plots.

Method clone()

The objects of this class are cloneable with this method.

Usage

QQDataMapping$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.