ThemePlotConfigurations

R6 class defining theme of plot configuration objects

Public fields

addScatter

theme properties for PlotConfiguration objects as used in function addScatter()

addLine

theme properties for PlotConfiguration objects as used in function addLine()

addRibbon

theme properties for PlotConfiguration objects as used in function addRibbon()

addErrorbar

theme properties for PlotConfiguration objects as used in function addErrorbar()

plotPKRatio

theme properties for PlotConfiguration objects as used in function plotPKRatio()

plotDDIRatio

theme properties for PlotConfiguration objects as used in function plotDDIRatio()

plotTimeProfile

theme properties for PlotConfiguration objects as used in function plotTimeProfile()

plotObsVsPred

theme properties for PlotConfiguration objects as used in function plotObsVsPred()

plotBoxWhisker

theme properties for PlotConfiguration objects as used in function plotBoxWhisker()

plotTornado

theme properties for PlotConfiguration objects as used in function plotTornado()

plotHistogram

theme properties for PlotConfiguration objects as used in function plotHistogram()

Methods

Public methods

ThemePlotConfigurations$new()

ThemePlotConfigurations$toJson()

ThemePlotConfigurations$clone()

Method new()

Create a new ThemePlotConfigurations object

Usage

ThemePlotConfigurations$new(
  addScatter = NULL,
  addLine = NULL,
  addRibbon = NULL,
  addErrorbar = NULL,
  plotPKRatio = NULL,
  plotDDIRatio = NULL,
  plotTimeProfile = NULL,
  plotObsVsPred = NULL,
  plotBoxWhisker = NULL,
  plotTornado = NULL,
  plotHistogram = NULL
)

Arguments

addScatter

theme properties for PlotConfiguration objects as used in function addScatter()

addLine

theme properties for PlotConfiguration objects as used in function addLine()

addRibbon

theme properties for PlotConfiguration objects as used in function addRibbon()

addErrorbar

theme properties for PlotConfiguration objects as used in function addErrorbar()

plotPKRatio

theme properties for PlotConfiguration objects as used in function plotPKRatio()

plotDDIRatio

theme properties for PlotConfiguration objects as used in function plotDDIRatio()

plotTimeProfile

theme properties for PlotConfiguration objects as used in function plotTimeProfile()

plotObsVsPred

theme properties for PlotConfiguration objects as used in function plotObsVsPred()

plotBoxWhisker

theme properties for PlotConfiguration objects as used in function plotBoxWhisker()

plotTornado

theme properties for PlotConfiguration objects as used in function plotTornado()

plotHistogram

theme properties for PlotConfiguration objects as used in function plotHistogram()

Returns

A new ThemePlotConfigurations object

Method toJson()

Translate object into a json list

Usage

ThemePlotConfigurations$toJson()

Returns

A list that can be saved into a json file

Method clone()

The objects of this class are cloneable with this method.

Usage

ThemePlotConfigurations$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.