TornadoPlotConfiguration

R6 class defining the configuration of a ggplot object for tornado plots

See also

Super class

tlf::PlotConfiguration -> TornadoPlotConfiguration

Public fields

bar

logical setting if tornado is uses a bar plot instead of regular points

colorPalette

color palette property from ggplot2

dodge

space between the bars/points

defaultYScale

Default yAxis scale value when creating a TornadoPlotConfiguration object

Methods

Public methods

TornadoPlotConfiguration$new()

TornadoPlotConfiguration$clone()

Method new()

Create a new TornadoPlotConfiguration object

Usage

TornadoPlotConfiguration$new(bar = TRUE, colorPalette = NULL, dodge = 0.5, ...)

Arguments

bar

logical setting if tornado is uses a bar plot instead of regular points

colorPalette

color palette property from ggplot2

dodge

space between the bars/points

...

parameters inherited from PlotConfiguration

Returns

A new TornadoPlotConfiguration object

Method clone()

The objects of this class are cloneable with this method.

Usage

TornadoPlotConfiguration$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.