addScatter

Add a scatter plot layer to a ggplot graphical object. Use optional argument caption to set legend caption. Since ggplot manage aesthetic properties across all layers, aesthetic properties defined in plotConfiguration will apply across all layers.

addScatter(
  data = NULL,
  metaData = NULL,
  x = NULL,
  y = NULL,
  caption = NULL,
  color = NULL,
  shape = NULL,
  size = NULL,
  linetype = NULL,
  dataMapping = NULL,
  plotConfiguration = NULL,
  plotObject = NULL
)

Arguments

data
data.frame containing the scatter points to be plotted

metaData
list of information on data such as dimension and unit of their variables

x
Mapping for x values. If data is NULL or not input, x numeric values will be used as is for the plot.

y
Mapping for y values. If data is NULL or not input, y numeric values will be used as is for the plot.

caption
vector of character strings defining the legend captions. This parameter is optional: default value NULL creates caption labels "data 1", "data 2" ...

color
vector of character strings defining the color of the scatter points. This parameter is optional: default value NULL will choose colors according to the current theme.

shape
vector of character strings or numerical defining the shapes of the scatter points. This parameter is optional: default value NULL will choose shapes according to the current theme.

size
vector of numerical defining the sizes of the scatter points. This parameter is optional: default value NULL will choose sizes according to the current theme.

linetype
vector of character strings defining the linetype linking the scatter points. This parameter is optional: default value NULL won't provide lines.

dataMapping
XYGDataMapping class or subclass object mapping x, y and aesthetic variables to the variable names of data.

plotConfiguration
PlotConfiguration object defining the labels, axes, background and legend properties of the plot.

plotObject
ggplot graphical object to which the line layer is added This parameter is optional: the tlf library will initialize an empty plot if the parameter is NULL or not provided

Value

A ggplot graphical object

Examples

# Add scatter using x and y
p <- addScatter(x = c(1, 2, 1, 2, 3), y = c(5, 0, 2, 3, 4))

# Add a custom scatter
time <- seq(0, 30, 0.1)
customScatterData <- data.frame(x = time, y = cos(time))

p <- addScatter(
  data = customScatterData,
  dataMapping = XYGDataMapping$new(x = "x", y = "y")
)

# Or for simple cases a smart mapping will get directly x and y from data
p <- addScatter(data = customScatterData)

# Add a scatter with caption
p <- addScatter(data = customScatterData, caption = "My scatter plot")

# Add a scatter with specific properties
p <- addScatter(data = customScatterData, color = "blue", shape = 19, size = 2, caption = "My data")

# Add a scatter with specific properties
pp <- addScatter(x = c(0, 1), y = c(1, 0), color = "red", shape = 20, size = 3, plotObject = p)