isBetween

Assess if x is between left and right bounds. Shortcut for x >= left & x <= right if strict=FALSE (default). Shortcut for x > left & x < right if strict=TRUE.

Usage

isBetween(x, left, right, strict = FALSE)

Arguments

x

Numeric values to assess

left

Numeric value(s) used as lower bound

right

Numeric value(s) used as upper bound

strict

Logical value defining if x is strictly between left and right. Default value is FALSE.

Value

Logical values

Examples

isBetween(1:12, 7, 9)
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE  TRUE FALSE FALSE FALSE

x <- rnorm(1e2)
x[isBetween(x, -1, 1)]
#>  [1] -0.481748243  0.536492849  0.046222113  0.737587386  0.856916405
#>  [6]  0.228807496  0.007155720 -0.625773846 -0.038260852 -0.942974216
#> [11] -0.014568023 -0.310487193  0.003670570 -0.649332295 -0.634342399
#> [16] -0.880161818  0.761150141  0.387562840 -0.028470060 -0.753239905
#> [21]  0.006173460 -0.032764618  0.830893543  0.377300142  0.970882809
#> [26] -0.239591615 -0.733655930 -0.764997542 -0.487501523  0.718448717
#> [31]  0.424405010 -0.907500422 -0.420015796 -0.334738068  0.330750736
#> [36] -0.861168075  0.072596903  0.977697502  0.499699137  0.381193005
#> [41]  0.820456028  0.861440069  0.217378263 -0.370887186  0.345816421
#> [46] -0.361594886 -0.117003586  0.627481136  0.958272311  0.856228918
#> [51]  0.113194951 -0.090169305  0.514184508 -0.981196100  0.513691355
#> [56] -0.821121267  0.482473672 -0.516502426  0.004565453

isBetween(x, cos(x) + 1, cos(x) - 1)
#>   [1] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [13] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [25] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [37] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [49] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [61] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [73] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [85] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [97] FALSE FALSE FALSE FALSE