plotPKRatio

Add PK Ratio plot layers to a ggplot graphical object. PK Ratio limits are plotted as horizontal lines. PK Ratios are plotted as a scatter plot.

plotPKRatio(
  data,
  metaData = NULL,
  dataMapping = NULL,
  plotConfiguration = NULL,
  plotObject = NULL
)

Arguments

data
data.frame containing the data to be used for the plot

metaData
list of lists containing complementary information to data (e.g. their unit and dimension). This parameter is optional.

dataMapping
A PKRatioDataMapping class or subclass mapping x and y variables to data variable names. dataMapping provides also the values of the PK Ratio limits plotted as horizontal lines. This parameter is optional: the tlf library provides a smart mapping if only data is provided and default values of the PK Ratio limits.

plotConfiguration
A PKRatioPlotConfiguration class or subclass defining labels, grid, background and watermark This parameter is optional: the tlf library provides a default configuration according to the current theme

plotObject
ggplot graphical object to which the PK Ratio plot layer is added This parameter is optional: the tlf library will initialize an empty plot if the parameter is NULL or not provided.

Value

A ggplot graphical object

References

For examples, see: https://www.open-systems-pharmacology.org/TLF-Library/articles/pk-ratio-vignette.html