Class for creating a plot grid

R6 class defining the configuration for {patchwork} plot grid used to create a grid of plots from {tlf}. It holds values for all relevant plot properties.

Value

A PlotGridConfiguration object.

Public fields

plotList

A list containing ggplot objects.

title, subtitle, caption

Text strings to use for the various plot annotations, where plot refers to the grid of plots as a whole.

tagLevels

A character vector defining the enumeration format to use at each level. Possible values are 'a' for lowercase letters, 'A' for uppercase letters, '1' for numbers, 'i' for lowercase Roman numerals, and 'I' for uppercase Roman numerals. It can also be a list containing character vectors defining arbitrary tag sequences. If any element in the list is a scalar and one of 'a', 'A', '1', 'i, or 'I', this level will be expanded to the expected sequence.

tagPrefix, tagSuffix

Strings that should appear before or after the tag.

tagSeparator

A separator between different tag levels

theme

A ggplot theme specification to use for the plot. Only elements related to the titles as well as plot margin and background will be used.

nColumns, nRows

The dimensions of the grid to create - if both are NULL it will use the same logic as facet_wrap() to set the dimensions

byRow

Analogous to byrow in matrix(). If FALSE the plots will be filled in in column-major order.

widths, heights

The relative widths and heights of each column and row in the grid. Will get repeated to match the dimensions of the grid.

guides

A string specifying how guides should be treated in the layout. 'collect' will collect guides below to the given nesting level, removing duplicates. 'keep' will stop collection at this level and let guides be placed alongside their plot. auto will allow guides to be collected if a upper level tries, but place them alongside the plot if not. If you modify default guide "position" with theme(legend.position=...) while also collecting guides you must apply that change to the overall patchwork.

design

Specification of the location of areas in the layout. Can either be specified as a text string or by concatenating calls to area() together. See the examples for further information on use.

Active bindings

title, subtitle, caption

Text strings to use for the various plot annotations, where plot refers to the grid of plots as a whole.

tagPrefix, tagSuffix

Strings that should appear before or after the tag.

nColumns, nRows

The dimensions of the grid to create - if both are NULL it will use the same logic as facet_wrap() to set the dimensions

widths, heights

The relative widths and heights of each column and row in the grid. Will get repeated to match the dimensions of the grid.

Methods

Public methods

PlotGridConfiguration$new()

PlotGridConfiguration$addPlots()

PlotGridConfiguration$clone()

Method new()

Create an instance of PlotGridConfiguration class.

Usage

PlotGridConfiguration$new(plotList = NULL)

Arguments

plotList

A list containing ggplot objects.

Returns

A PlotGridConfiguration object.

Method addPlots()

Add a plot object.

Usage

PlotGridConfiguration$addPlots(plots = NULL)

Arguments

plots

A single or a list containing ggplot object(s).

Returns

PlotGridConfiguration object with $plotList field updated to store entered plots.

Examples

library(ggplot2)

myPlotGrid <- PlotGridConfiguration$new()

# You can add a single ggplot object
p <- ggplot(mtcars, aes(wt, mpg)) + geom_point()
myPlotGrid$addPlots(p)

# Or you can also pass a list
myPlotGrid$addPlots(list("p1" = ggplot(), "p2" = ggplot()))

# Since we added three plots, the `plotList` field should
# now be a list of length `3`
length(myPlotGrid$plotList)

Method clone()

The objects of this class are cloneable with this method.

Usage

PlotGridConfiguration$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `PlotGridConfiguration$addPlots`
## ------------------------------------------------

library(ggplot2)

myPlotGrid <- PlotGridConfiguration$new()

# You can add a single ggplot object
p <- ggplot(mtcars, aes(wt, mpg)) + geom_point()
myPlotGrid$addPlots(p)

# Or you can also pass a list
myPlotGrid$addPlots(list("p1" = ggplot(), "p2" = ggplot()))

# Since we added three plots, the `plotList` field should
# now be a list of length `3`
length(myPlotGrid$plotList)
#> [1] 3