getGuestValues

Get a data.frame with Guest et al. ratio limits with:

ymax = x.limit

ymin = x/limit

limit = (delta+2(x-1))/x

getGuestValues(x, delta = 1, residualsVsObserved = FALSE)

Arguments

x
Numeric values input of Guest function

delta
Numeric value parameter of Guest function

residualsVsObserved
Logical value defining if limits are claculated as residuals vs observed, instead of predicted vs observed.

Value

A data.frame with x, ymin and ymax defining Guest et al. limits

References

https://dmd.aspetjournals.org/content/39/2/170

Examples

# Get predicted vs observed Guest et al. limits
getGuestValues(x = 10^seq(-2, 2, 0.2))
#>               x         ymin         ymax
#> 1    0.01000000  0.005025126   0.01990000
#> 2    0.01584893  0.007987765   0.03144668
#> 3    0.02511886  0.012719178   0.04960677
#> 4    0.03981072  0.020309629   0.07803654
#> 5    0.06309573  0.032575557   0.12221040
#> 6    0.10000000  0.052631579   0.19000000
#> 7    0.15848932  0.086064838   0.29185977
#> 8    0.25118864  0.143633927   0.43928155
#> 9    0.39810717  0.248522974   0.63772502
#> 10   0.63095734  0.460874862   0.86380752
#> 11   1.00000000  1.000000000   1.00000000
#> 12   1.58489319  1.157665312   2.16978638
#> 13   2.51188643  1.568073959   4.02377286
#> 14   3.98107171  2.276444335   6.96214341
#> 15   6.30957344  3.426302932  11.61914689
#> 16  10.00000000  5.263157895  19.00000000
#> 17  15.84893192  8.182609851  30.69786385
#> 18  25.11886432 12.814509565  49.23772863
#> 19  39.81071706 20.158538322  78.62143411
#> 20  63.09573445 31.799864165 125.19146890
#> 21 100.00000000 50.251256281 199.00000000

# Get residuals vs observed Guest et al. limits
getGuestValues(x = 10^seq(-2, 2, 0.2), residualsVsObserved = TRUE)
#>               x      ymin     ymax
#> 1    0.01000000 0.5025126 1.990000
#> 2    0.01584893 0.5039939 1.984151
#> 3    0.02511886 0.5063596 1.974881
#> 4    0.03981072 0.5101548 1.960189
#> 5    0.06309573 0.5162878 1.936904
#> 6    0.10000000 0.5263158 1.900000
#> 7    0.15848932 0.5430324 1.841511
#> 8    0.25118864 0.5718170 1.748811
#> 9    0.39810717 0.6242615 1.601893
#> 10   0.63095734 0.7304374 1.369043
#> 11   1.00000000 1.0000000 1.000000
#> 12   1.58489319 0.7304374 1.369043
#> 13   2.51188643 0.6242615 1.601893
#> 14   3.98107171 0.5718170 1.748811
#> 15   6.30957344 0.5430324 1.841511
#> 16  10.00000000 0.5263158 1.900000
#> 17  15.84893192 0.5162878 1.936904
#> 18  25.11886432 0.5101548 1.960189
#> 19  39.81071706 0.5063596 1.974881
#> 20  63.09573445 0.5039939 1.984151
#> 21 100.00000000 0.5025126 1.990000