getGuestValuesFromDataMapping

Get a data.frame with Guest et al. ratio limits from data and its DDIRatioDataMapping

Usage

getGuestValuesFromDataMapping(data, dataMapping)

Arguments

data

A data.frame to use for plot.

dataMapping

A DDIRatioDataMapping object mapping x, y and aesthetic groups to their variable names of data.

Value

A data.frame with x, ymin and ymax defining Guest et al. limits

Examples

# Get the data.frame of Guest et al. limits
ddiData <- data.frame(x = c(1, 2, 1, 2, 3), y = c(5, 0.2, 2, 3, 4))

getGuestValuesFromDataMapping(
  data = ddiData,
  dataMapping = DDIRatioDataMapping$new(x = "x", y = "y")
)
#>                x         ymin         ymax
#> 1     0.01000000  0.005025126   0.01990000
#> 2     0.01018629  0.005119218   0.02026882
#> 3     0.01037605  0.005215081   0.02064444
#> 4     0.01056935  0.005312749   0.02102698
#> 5     0.01076624  0.005412256   0.02141657
#> 6     0.01096681  0.005513636   0.02181334
#> 7     0.01117111  0.005616927   0.02221742
#> 8     0.01137921  0.005722163   0.02262894
#> 9     0.01159120  0.005829383   0.02304804
#> 10    0.01180713  0.005938623   0.02347485
#> 11    0.01202708  0.006049923   0.02390952
#> 12    0.01225114  0.006163322   0.02435218
#> 13    0.01247936  0.006278859   0.02480299
#> 14    0.01271184  0.006396576   0.02526209
#> 15    0.01294865  0.006516514   0.02572963
#> 16    0.01318987  0.006638716   0.02620577
#> 17    0.01343558  0.006763225   0.02669065
#> 18    0.01368587  0.006890086   0.02718445
#> 19    0.01394083  0.007019342   0.02768731
#> 20    0.01420053  0.007151040   0.02819941
#> 21    0.01446507  0.007285227   0.02872091
#> 22    0.01473454  0.007421951   0.02925198
#> 23    0.01500903  0.007561260   0.02979279
#> 24    0.01528864  0.007703204   0.03034353
#> 25    0.01557345  0.007847833   0.03090436
#> 26    0.01586357  0.007995199   0.03147548
#> 27    0.01615909  0.008145355   0.03205706
#> 28    0.01646012  0.008298353   0.03264929
#> 29    0.01676675  0.008454250   0.03325238
#> 30    0.01707910  0.008613101   0.03386650
#> 31    0.01739726  0.008774962   0.03449186
#> 32    0.01772136  0.008939892   0.03512867
#> 33    0.01805149  0.009107950   0.03577712
#> 34    0.01838777  0.009279197   0.03643743
#> 35    0.01873032  0.009453693   0.03710981
#> 36    0.01907924  0.009631502   0.03779447
#> 37    0.01943467  0.009812688   0.03849163
#> 38    0.01979672  0.009997316   0.03920153
#> 39    0.02016551  0.010185453   0.03992437
#> 40    0.02054117  0.010377166   0.04066041
#> 41    0.02092383  0.010572526   0.04140986
#> 42    0.02131362  0.010771603   0.04217298
#> 43    0.02171068  0.010974470   0.04295000
#> 44    0.02211512  0.011181199   0.04374117
#> 45    0.02252711  0.011391866   0.04454674
#> 46    0.02294676  0.011606548   0.04536697
#> 47    0.02337424  0.011825323   0.04620212
#> 48    0.02380968  0.012048271   0.04705245
#> 49    0.02425323  0.012275474   0.04791824
#> 50    0.02470504  0.012507013   0.04879974
#> 51    0.02516527  0.012742975   0.04969725
#> 52    0.02563407  0.012983446   0.05061104
#> 53    0.02611161  0.013228514   0.05154141
#> 54    0.02659804  0.013478270   0.05248863
#> 55    0.02709354  0.013732804   0.05345302
#> 56    0.02759826  0.013992212   0.05443486
#> 57    0.02811239  0.014256589   0.05543448
#> 58    0.02863610  0.014526033   0.05645217
#> 59    0.02916956  0.014800644   0.05748825
#> 60    0.02971296  0.015080523   0.05854306
#> 61    0.03026648  0.015365774   0.05961690
#> 62    0.03083031  0.015656505   0.06071012
#> 63    0.03140465  0.015952823   0.06182305
#> 64    0.03198969  0.016254838   0.06295604
#> 65    0.03258562  0.016562665   0.06410943
#> 66    0.03319266  0.016876418   0.06528357
#> 67    0.03381101  0.017196215   0.06647883
#> 68    0.03444087  0.017522176   0.06769557
#> 69    0.03508247  0.017854424   0.06893416
#> 70    0.03573602  0.018193085   0.07019498
#> 71    0.03640175  0.018538287   0.07147841
#> 72    0.03707988  0.018890161   0.07278484
#> 73    0.03777064  0.019248839   0.07411465
#> 74    0.03847427  0.019614459   0.07546826
#> 75    0.03919100  0.019987160   0.07684607
#> 76    0.03992109  0.020367084   0.07824849
#> 77    0.04066478  0.020754377   0.07967594
#> 78    0.04142232  0.021149187   0.08112884
#> 79    0.04219398  0.021551666   0.08260763
#> 80    0.04298001  0.021961969   0.08411274
#> 81    0.04378069  0.022380255   0.08564462
#> 82    0.04459628  0.022806685   0.08720372
#> 83    0.04542706  0.023241425   0.08879050
#> 84    0.04627332  0.023684644   0.09040542
#> 85    0.04713535  0.024136514   0.09204895
#> 86    0.04801343  0.024597213   0.09372157
#> 87    0.04890787  0.025066920   0.09542376
#> 88    0.04981898  0.025545821   0.09715602
#> 89    0.05074705  0.026034103   0.09891884
#> 90    0.05169242  0.026531960   0.10071273
#> 91    0.05265540  0.027039588   0.10253820
#> 92    0.05363631  0.027557190   0.10439577
#> 93    0.05463550  0.028084970   0.10628597
#> 94    0.05565331  0.028623140   0.10820932
#> 95    0.05669007  0.029171915   0.11016638
#> 96    0.05774615  0.029731516   0.11215769
#> 97    0.05882190  0.030302168   0.11418379
#> 98    0.05991770  0.030884100   0.11624526
#> 99    0.06103390  0.031477550   0.11834267
#> 100   0.06217090  0.032082758   0.12047658
#> 101   0.06332908  0.032699972   0.12264760
#> 102   0.06450884  0.033329442   0.12485629
#> 103   0.06571058  0.033971429   0.12710327
#> 104   0.06693470  0.034626196   0.12938914
#> 105   0.06818162  0.035294013   0.13171451
#> 106   0.06945178  0.035975158   0.13408001
#> 107   0.07074559  0.036669915   0.13648625
#> 108   0.07206351  0.037378572   0.13893388
#> 109   0.07340598  0.038101428   0.14142353
#> 110   0.07477346  0.038838787   0.14395586
#> 111   0.07616642  0.039590959   0.14653151
#> 112   0.07758532  0.040358265   0.14915116
#> 113   0.07903066  0.041141030   0.15181547
#> 114   0.08050292  0.041939589   0.15452512
#> 115   0.08200261  0.042754284   0.15728078
#> 116   0.08353023  0.043585468   0.16008316
#> 117   0.08508632  0.044433499   0.16293295
#> 118   0.08667139  0.045298747   0.16583085
#> 119   0.08828599  0.046181588   0.16877756
#> 120   0.08993067  0.047082410   0.17177381
#> 121   0.09160598  0.048001610   0.17482031
#> 122   0.09331251  0.048939594   0.17791780
#> 123   0.09505083  0.049896780   0.18106700
#> 124   0.09682153  0.050873595   0.18426865
#> 125   0.09862522  0.051870478   0.18752350
#> 126   0.10046251  0.052887877   0.19083230
#> 127   0.10233402  0.053926256   0.19419579
#> 128   0.10424040  0.054986086   0.19761474
#> 129   0.10618229  0.056067854   0.20108991
#> 130   0.10816036  0.057172057   0.20462206
#> 131   0.11017528  0.058299206   0.20821197
#> 132   0.11222774  0.059449828   0.21186041
#> 133   0.11431843  0.060624459   0.21556815
#> 134   0.11644806  0.061823653   0.21933597
#> 135   0.11861737  0.063047979   0.22316466
#> 136   0.12082709  0.064298018   0.22705500
#> 137   0.12307798  0.065574371   0.23100777
#> 138   0.12537080  0.066877652   0.23502376
#> 139   0.12770633  0.068208493   0.23910375
#> 140   0.13008537  0.069567545   0.24324854
#> 141   0.13250873  0.070955474   0.24745890
#> 142   0.13497723  0.072372968   0.25173561
#> 143   0.13749172  0.073820731   0.25607947
#> 144   0.14005305  0.075299489   0.26049125
#> 145   0.14266210  0.076809988   0.26497173
#> 146   0.14531975  0.078352995   0.26952167
#> 147   0.14802691  0.079929300   0.27414186
#> 148   0.15078450  0.081539715   0.27883304
#> 149   0.15359347  0.083185076   0.28359598
#> 150   0.15645476  0.084866243   0.28843143
#> 151   0.15936935  0.086584103   0.29334012
#> 152   0.16233824  0.088339567   0.29832278
#> 153   0.16536244  0.090133574   0.30338014
#> 154   0.16844298  0.091967093   0.30851292
#> 155   0.17158090  0.093841121   0.31372179
#> 156   0.17477728  0.095756685   0.31900746
#> 157   0.17803320  0.097714844   0.32437058
#> 158   0.18134978  0.099716690   0.32981182
#> 159   0.18472814  0.101763349   0.33533180
#> 160   0.18816944  0.103855982   0.34093115
#> 161   0.19167485  0.105995788   0.34661045
#> 162   0.19524556  0.108184002   0.35237029
#> 163   0.19888279  0.110421900   0.35821121
#> 164   0.20258777  0.112710801   0.36413374
#> 165   0.20636178  0.115052062   0.37013837
#> 166   0.21020609  0.117447091   0.37622558
#> 167   0.21412202  0.119897338   0.38239579
#> 168   0.21811089  0.122404302   0.38864942
#> 169   0.22217408  0.124969535   0.39498684
#> 170   0.22631296  0.127594638   0.40140836
#> 171   0.23052894  0.130281271   0.40791429
#> 172   0.23482346  0.133031147   0.41450486
#> 173   0.23919798  0.135846042   0.42118029
#> 174   0.24365400  0.138727791   0.42794073
#> 175   0.24819303  0.141678297   0.43478628
#> 176   0.25281661  0.144699530   0.44171699
#> 177   0.25752633  0.147793529   0.44873285
#> 178   0.26232379  0.150962410   0.45583381
#> 179   0.26721062  0.154208364   0.46301972
#> 180   0.27218848  0.157533664   0.47029039
#> 181   0.27725908  0.160940669   0.47764556
#> 182   0.28242413  0.164431824   0.48508487
#> 183   0.28768541  0.168009670   0.49260792
#> 184   0.29304470  0.171676843   0.50021420
#> 185   0.29850382  0.175436083   0.50790312
#> 186   0.30406465  0.179290236   0.51567399
#> 187   0.30972907  0.183242261   0.52352604
#> 188   0.31549901  0.187295233   0.53145839
#> 189   0.32137643  0.191452355   0.53947006
#> 190   0.32736335  0.195716956   0.54755994
#> 191   0.33346180  0.200092504   0.55572683
#> 192   0.33967386  0.204582612   0.56396939
#> 193   0.34600164  0.209191041   0.57228614
#> 194   0.35244730  0.213921716   0.58067550
#> 195   0.35901304  0.218778726   0.58913571
#> 196   0.36570109  0.223766341   0.59766489
#> 197   0.37251373  0.228889016   0.60626098
#> 198   0.37945328  0.234151402   0.61492177
#> 199   0.38652212  0.239558360   0.62364489
#> 200   0.39372263  0.245114973   0.63242775
#> 201   0.40105729  0.250826553   0.64126763
#> 202   0.40852858  0.256698660   0.65016156
#> 203   0.41613906  0.262737115   0.65910640
#> 204   0.42389131  0.268948016   0.66809877
#> 205   0.43178797  0.275337751   0.67713509
#> 206   0.43983175  0.281913022   0.68621153
#> 207   0.44802537  0.288680858   0.69532401
#> 208   0.45637163  0.295648640   0.70446819
#> 209   0.46487337  0.302824120   0.71363949
#> 210   0.47353349  0.310215448   0.72283302
#> 211   0.48235494  0.317831196   0.73204360
#> 212   0.49134073  0.325680383   0.74126575
#> 213   0.50049391  0.333772509   0.75049367
#> 214   0.50981761  0.342117588   0.75972122
#> 215   0.51931499  0.350726178   0.76894192
#> 216   0.52898931  0.359609424   0.77814893
#> 217   0.53884384  0.368779095   0.78733500
#> 218   0.54888196  0.378247631   0.79649251
#> 219   0.55910708  0.388028193   0.80561343
#> 220   0.56952268  0.398134713   0.81468928
#> 221   0.58013231  0.408581951   0.82371112
#> 222   0.59093959  0.419385560   0.83266958
#> 223   0.60194820  0.430562155   0.84155476
#> 224   0.61316188  0.442129386   0.85035627
#> 225   0.62458447  0.454106020   0.85906318
#> 226   0.63621985  0.466512032   0.86766400
#> 227   0.64807198  0.479368705   0.87614667
#> 228   0.66014491  0.492698736   0.88449852
#> 229   0.67244274  0.506526358   0.89270624
#> 230   0.68496967  0.520877471   0.90075589
#> 231   0.69772996  0.535779789   0.90863283
#> 232   0.71072797  0.551263002   0.91632169
#> 233   0.72396811  0.567358952   0.92380640
#> 234   0.73745491  0.584101839   0.93107008
#> 235   0.75119295  0.601528434   0.93809505
#> 236   0.76518692  0.619678334   0.94486282
#> 237   0.77944158  0.638594236   0.95135398
#> 238   0.79396178  0.658322244   0.95754825
#> 239   0.80875249  0.678912220   0.96342439
#> 240   0.82381873  0.700418172   0.96896016
#> 241   0.83916564  0.722898698   0.97413231
#> 242   0.85479845  0.746417480   0.97891651
#> 243   0.87072248  0.771043852   0.98328732
#> 244   0.88694317  0.796853438   0.98721815
#> 245   0.90346602  0.823928887   0.99068119
#> 246   0.92029668  0.852360704   0.99364738
#> 247   0.93744088  0.882248206   0.99608636
#> 248   0.95490446  0.913700629   0.99796639
#> 249   0.97269336  0.946838388   0.99925435
#> 250   0.99081366  0.981794555   0.99991561
#> 251   1.00927151  1.000084396   1.01854303
#> 252   1.02807322  1.000746209   1.05614645
#> 253   1.04722519  1.002037752   1.09445038
#> 254   1.06673394  1.003929020   1.13346788
#> 255   1.08660611  1.006393233   1.17321223
#> 256   1.10684849  1.009406466   1.21369698
#> 257   1.12746796  1.012947339   1.25493592
#> 258   1.14847155  1.016996737   1.29694310
#> 259   1.16986641  1.021537577   1.33973283
#> 260   1.19165984  1.026554596   1.38331969
#> 261   1.21385926  1.032034174   1.42771853
#> 262   1.23647224  1.037964173   1.47294447
#> 263   1.25950646  1.044333793   1.51901293
#> 264   1.28296980  1.051133458   1.56593960
#> 265   1.30687023  1.058354698   1.61374046
#> 266   1.33121590  1.065990060   1.66243181
#> 267   1.35601511  1.074033015   1.71203022
#> 268   1.38127630  1.082477890   1.76255261
#> 269   1.40700809  1.091319793   1.81401617
#> 270   1.43321923  1.100554560   1.86643845
#> 271   1.45991865  1.110178694   1.91983730
#> 272   1.48711546  1.120189321   1.97423093
#> 273   1.51481892  1.130584145   2.02963784
#> 274   1.54303847  1.141361411   2.08607694
#> 275   1.57178372  1.152519867   2.14356744
#> 276   1.60106446  1.164058736   2.20212892
#> 277   1.63089068  1.175977684   2.26178135
#> 278   1.66127252  1.188276800   2.32254504
#> 279   1.69222035  1.200956566   2.38444070
#> 280   1.72374471  1.214017842   2.44748942
#> 281   1.75585633  1.227461844   2.51171266
#> 282   1.78856616  1.241290129   2.57713232
#> 283   1.82188534  1.255504580   2.64377069
#> 284   1.85582523  1.270107388   2.71165046
#> 285   1.89039738  1.285101045   2.78079476
#> 286   1.92561357  1.300488331   2.85122714
#> 287   1.96148581  1.316272301   2.92297162
#> 288   1.99802631  1.332456282   2.99605262
#> 289   2.03524752  1.349043856   3.07049504
#> 290   2.07316213  1.366038863   3.14632425
#> 291   2.11178304  1.383445384   3.22356609
#> 292   2.15112343  1.401267745   3.30224686
#> 293   2.19119669  1.419510504   3.38239338
#> 294   2.23201647  1.438178449   3.46403294
#> 295   2.27359668  1.457276597   3.54719336
#> 296   2.31595149  1.476810184   3.63190299
#> 297   2.35909533  1.496784669   3.71819066
#> 298   2.40304289  1.517205726   3.80608579
#> 299   2.44780916  1.538079243   3.89561831
#> 300   2.49340937  1.559411323   3.98681874
#> 301   2.53985907  1.581208278   4.07971814
#> 302   2.58717408  1.603476629   4.17434816
#> 303   2.63537052  1.626223110   4.27074104
#> 304   2.68446481  1.649454659   4.36892962
#> 305   2.73447368  1.673178425   4.46894736
#> 306   2.78541416  1.697401763   4.57082833
#> 307   2.83730362  1.722132237   4.67460723
#> 308   2.89015972  1.747377619   4.78031944
#> 309   2.94400048  1.773145890   4.88800095
#> 310   2.99884423  1.799445242   4.99768846
#> 311   3.05470967  1.826284075   5.10941934
#> 312   3.11161583  1.853671003   5.22323166
#> 313   3.16958209  1.881614853   5.33916418
#> 314   3.22862820  1.910124666   5.45725641
#> 315   3.28877429  1.939209702   5.57754857
#> 316   3.35004083  1.968879436   5.70008166
#> 317   3.41244871  1.999143565   5.82489742
#> 318   3.47601918  2.030012008   5.95203836
#> 319   3.54077391  2.061494910   6.08154782
#> 320   3.60673495  2.093602643   6.21346990
#> 321   3.67392478  2.126345807   6.34784956
#> 322   3.74236629  2.159735236   6.48473258
#> 323   3.81208280  2.193782000   6.62416559
#> 324   3.88309805  2.228497407   6.76619610
#> 325   3.95543624  2.263893005   6.91087249
#> 326   4.02912203  2.299980588   7.05824406
#> 327   4.10418050  2.336772199   7.20836101
#> 328   4.18063724  2.374280131   7.36127448
#> 329   4.25851829  2.412516933   7.51703658
#> 330   4.33785019  2.451495412   7.67570038
#> 331   4.41865996  2.491228638   7.83731991
#> 332   4.50097513  2.531729948   8.00195026
#> 333   4.58482375  2.573012951   8.16964750
#> 334   4.67023439  2.615091529   8.34046878
#> 335   4.75723614  2.657979842   8.51447228
#> 336   4.84585864  2.701692338   8.69171729
#> 337   4.93613210  2.746243749   8.87226420
#> 338   5.02808725  2.791649103   9.05617451
#> 339   5.12175544  2.837923722   9.24351089
#> 340   5.21716858  2.885083235   9.43433715
#> 341   5.31435916  2.933143575   9.62871832
#> 342   5.41336030  2.982120989   9.82672061
#> 343   5.51420574  3.032032042  10.02841148
#> 344   5.61692982  3.082893623  10.23385965
#> 345   5.72156756  3.134722949  10.44313511
#> 346   5.82815458  3.187537571  10.65630916
#> 347   5.93672722  3.241355381  10.87345443
#> 348   6.04732245  3.296194618  11.09464490
#> 349   6.15997796  3.352073872  11.31995592
#> 350   6.27473213  3.409012091  11.54946426
#> 351   6.39162405  3.467028587  11.78324811
#> 352   6.51069356  3.526143047  12.02138711
#> 353   6.63198120  3.586375530  12.26396240
#> 354   6.75552832  3.647746483  12.51105663
#> 355   6.88137699  3.710276742  12.76275397
#> 356   7.00957009  3.773987543  13.01914018
#> 357   7.14015130  3.838900524  13.28030261
#> 358   7.27316511  3.905037739  13.54633023
#> 359   7.40865683  3.972421659  13.81731367
#> 360   7.54667263  4.041075184  14.09334526
#> 361   7.68725952  4.111021647  14.37451904
#> 362   7.83046540  4.182284825  14.66093081
#> 363   7.97633907  4.254888947  14.95267814
#> 364   8.12493021  4.328858698  15.24986042
#> 365   8.27628946  4.404219233  15.55257891
#> 366   8.43046837  4.480996181  15.86093674
#> 367   8.58751948  4.559215656  16.17503897
#> 368   8.74749630  4.638904265  16.49499260
#> 369   8.91045332  4.720089119  16.82090665
#> 370   9.07644607  4.802797838  17.15289215
#> 371   9.24553110  4.887058563  17.49106220
#> 372   9.41776601  4.972899967  17.83553201
#> 373   9.59320948  5.060351261  18.18641896
#> 374   9.77192128  5.149442206  18.54384257
#> 375   9.95396231  5.240203124  18.90792462
#> 376  10.13939458  5.332664908  19.27878915
#> 377  10.32828126  5.426859028  19.65656252
#> 378  10.52068671  5.522817550  20.04137342
#> 379  10.71667648  5.620573139  20.43335296
#> 380  10.91631734  5.720159073  20.83263468
#> 381  11.11967731  5.821609258  21.23935462
#> 382  11.32682567  5.924958233  21.65365134
#> 383  11.53783300  6.030241185  22.07566599
#> 384  11.75277117  6.137493962  22.50554235
#> 385  11.97171343  6.246753084  22.94342687
#> 386  12.19473437  6.358055755  23.38946873
#> 387  12.42190995  6.471439874  23.84381991
#> 388  12.65331759  6.586944054  24.30663519
#> 389  12.88903612  6.704607628  24.77807225
#> 390  13.12914585  6.824470666  25.25829170
#> 391  13.37372858  6.946573988  25.74745716
#> 392  13.62286764  7.070959178  26.24573528
#> 393  13.87664791  7.197668598  26.75329582
#> 394  14.13515585  7.326745403  27.27031170
#> 395  14.39847953  7.458233554  27.79695906
#> 396  14.66670866  7.592177835  28.33341733
#> 397  14.93993464  7.728623867  28.87986927
#> 398  15.21825053  7.867618124  29.43650107
#> 399  15.50175117  8.009207947  30.00350235
#> 400  15.79053314  8.153441565  30.58106629
#> 401  16.08469483  8.300368106  31.16938967
#> 402  16.38433646  8.450037615  31.76867291
#> 403  16.68956010  8.602501075  32.37912020
#> 404  17.00046975  8.757810418  33.00093950
#> 405  17.31717134  8.916018548  33.63434267
#> 406  17.63977275  9.077179355  34.27954551
#> 407  17.96838391  9.241347739  34.93676782
#> 408  18.30311676  9.408579621  35.60623351
#> 409  18.64408534  9.578931967  36.28817068
#> 410  18.99140582  9.752462809  36.98281165
#> 411  19.34519654  9.929231259  37.69039307
#> 412  19.70557802 10.109297533  38.41115603
#> 413  20.07267304 10.292722973  39.14534607
#> 414  20.44660667 10.479570063  39.89321333
#> 415  20.82750630 10.669902454  40.65501260
#> 416  21.21550171 10.863784985  41.43100343
#> 417  21.61072509 11.061283705  42.22145018
#> 418  22.01331107 11.262465894  43.02662215
#> 419  22.42339683 11.467400087  43.84679366
#> 420  22.84112207 11.676156100  44.68224415
#> 421  23.26662911 11.888805049  45.53325823
#> 422  23.70006292 12.105419376  46.40012584
#> 423  24.14157116 12.326072877  47.28314232
#> 424  24.59130425 12.550840721  48.18260851
#> 425  25.04941542 12.779799482  49.09883084
#> 426  25.51606074 13.013027159  50.03212147
#> 427  25.99139918 13.250603207  50.98279837
#> 428  26.47559271 13.492608563  51.95118541
#> 429  26.96880626 13.739125672  52.93761253
#> 430  27.47120789 13.990238518  53.94241579
#> 431  27.98296876 14.246032650  54.96593751
#> 432  28.50426321 14.506595210  56.00852642
#> 433  29.03526885 14.772014969  57.07053770
#> 434  29.57616659 15.042382349  58.15233318
#> 435  30.12714071 15.317789459  59.25428142
#> 436  30.68837892 15.598330126  60.37675784
#> 437  31.26007243 15.884099925  61.52014486
#> 438  31.84241602 16.175196213  62.68483203
#> 439  32.43560807 16.471718162  63.87121614
#> 440  33.03985070 16.773766793  65.07970140
#> 441  33.65534976 17.081445008  66.31069951
#> 442  34.28231494 17.394857630  67.56462987
#> 443  34.92095985 17.714111432  68.84191970
#> 444  35.57150207 18.039315181  70.14300414
#> 445  36.23416323 18.370579670  71.46832647
#> 446  36.90916911 18.708017754  72.81833821
#> 447  37.59674966 19.051744396  74.19349932
#> 448  38.29713914 19.401876699  75.59427828
#> 449  39.01057617 19.758533948  77.02115234
#> 450  39.73730381 20.121837651  78.47460763
#> 451  40.47756966 20.491911583  79.95513932
#> 452  41.23162591 20.868881824  81.46325182
#> 453  41.99972947 21.252876801  82.99945893
#> 454  42.78214202 21.644027340  84.56428404
#> 455  43.57913012 22.042466699  86.15826025
#> 456  44.39096531 22.448330623  87.78193063
#> 457  45.21792417 22.861757387  89.43584835
#> 458  46.06028844 23.282887838  91.12057688
#> 459  46.91834511 23.711865454  92.83669021
#> 460  47.79238650 24.148836381  94.58477300
#> 461  48.68271040 24.593949492  96.36542080
#> 462  49.58962014 25.047356432  98.17924028
#> 463  50.51342469 25.509211673 100.02684938
#> 464  51.45443878 25.979672564 101.90887757
#> 465  52.41298302 26.458899387 103.82596604
#> 466  53.38938397 26.947055409 105.77876794
#> 467  54.38397428 27.444306942 107.76794857
#> 468  55.39709281 27.950823393 109.79418562
#> 469  56.42908471 28.466777329 111.85816943
#> 470  57.48030158 28.992344531 113.96060316
#> 471  58.55110156 29.527704055 116.10220312
#> 472  59.64184946 30.073038292 118.28369892
#> 473  60.75291689 30.628533033 120.50583378
#> 474  61.88468239 31.194377532 122.76936477
#> 475  63.03753153 31.770764565 125.07506306
#> 476  64.21185709 32.357890504 127.42371418
#> 477  65.40805915 32.955955376 129.81611830
#> 478  66.62654525 33.565162938 132.25309049
#> 479  67.86773051 34.185720741 134.73546101
#> 480  69.13203780 34.817840205 137.26407559
#> 481  70.41989785 35.461736687 139.83979571
#> 482  71.73174944 36.117629558 142.46349889
#> 483  73.06803951 36.785742274 145.13607901
#> 484  74.42922330 37.466302457 147.85844660
#> 485  75.81576458 38.159541967 150.63152915
#> 486  77.22813571 38.865696986 153.45627143
#> 487  78.66681790 39.585008094 156.33363580
#> 488  80.13230128 40.317720357 159.26460257
#> 489  81.62508514 41.064083402 162.25017029
#> 490  83.14567806 41.824351510 165.29135612
#> 491  84.69459808 42.598783697 168.38919617
#> 492  86.27237292 43.387643808 171.54474585
#> 493  87.87954012 44.191200600 174.75908024
#> 494  89.51664721 45.009727839 178.03329443
#> 495  91.18425196 45.843504390 181.36850392
#> 496  92.88292250 46.692814315 184.76584500
#> 497  94.61323756 47.557946967 188.22647512
#> 498  96.37578664 48.439197089 191.75157328
#> 499  98.17117023 49.336864918 195.34234046
#> 500 100.00000000 50.251256281 199.00000000