isBetween

Assess if x is between left and right bounds. Shortcut for x >= left & x <= right if strict=FALSE (default). Shortcut for x > left & x < right if strict=TRUE.

Usage

isBetween(x, left, right, strict = FALSE)

Arguments

x

Numeric values to assess

left

Numeric value(s) used as lower bound

right

Numeric value(s) used as upper bound

strict

Logical value defining if x is strictly between left and right. Default value is FALSE.

Value

Logical values

Examples

isBetween(1:12, 7, 9)
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE  TRUE FALSE FALSE FALSE

x <- rnorm(1e2)
x[isBetween(x, -1, 1)]
#>  [1] -0.08013185 -0.03228341 -0.71898093 -0.78026990 -0.42783487  0.03398894
#>  [7]  0.74079278 -0.16356423 -0.61094549  0.25031947 -0.85541313 -0.14490163
#> [13] -0.32444696 -0.17256490 -0.09450402  0.03255579  0.46129012 -0.41647627
#> [19]  0.68094267 -0.41437304 -0.51834551 -0.68401973 -0.88564860  0.04923707
#> [25]  0.18556122 -0.60865779 -0.73110285 -0.64601525 -0.93245461 -0.76908693
#> [31]  0.37157978  0.35543278 -0.98399849  0.21472959 -0.08008508 -0.44025034
#> [37] -0.77816901  0.77750668 -0.18358770 -0.21324238  0.93053526  0.41081180
#> [43] -0.78236663 -0.12639591 -0.74329990 -0.30422384  0.33765806 -0.60750209
#> [49] -0.29556027 -0.13453714  0.81478437 -0.27292173  0.74338485  0.32781805
#> [55] -0.53416511  0.02893134 -0.39567707 -0.69486833 -0.34701739 -0.04025917
#> [61] -0.57390128 -0.70595925 -0.24697048

isBetween(x, cos(x) + 1, cos(x) - 1)
#>   [1] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [13] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [25] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [37] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [49] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [61] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [73] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [85] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [97] FALSE FALSE FALSE FALSE