isBetween

Assess if x is between left and right bounds. Shortcut for x >= left & x <= right if strict=FALSE (default). Shortcut for x > left & x < right if strict=TRUE.

Usage

isBetween(x, left, right, strict = FALSE)

Arguments

x

Numeric values to assess

left

Numeric value(s) used as lower bound

right

Numeric value(s) used as upper bound

strict

Logical value defining if x is strictly between left and right. Default value is FALSE.

Value

Logical values

Examples

isBetween(1:12, 7, 9)
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE  TRUE FALSE FALSE FALSE

x <- rnorm(1e2)
x[isBetween(x, -1, 1)]
#>  [1] -0.86860426  0.27810480 -0.91195985  0.42540108  0.58508968  0.09767160
#>  [7]  0.34155225 -0.87540838 -0.30500367  0.06017962  0.14302714 -0.73624415
#> [13]  0.07339946 -0.50804975  0.53625657 -0.77502802  0.90247243 -0.44106857
#> [19] -0.08786845  0.65090230 -0.56702404 -0.80859542 -0.19159999 -0.94930198
#> [25]  0.13435945 -0.08982843  0.41576340 -0.05276224  0.67516335  0.38181668
#> [31] -0.96506232 -0.37532620 -0.70251503 -0.19816324  0.85867043 -0.80561835
#> [37] -0.08566294 -0.72395446  0.94703290  0.14698074  0.45288572  0.99393774
#> [43] -0.84680562  0.46451027  0.64312414  0.90714054 -0.05273945 -0.66398242
#> [49] -0.29244573 -0.47635846 -0.73297830  0.98458437  0.73620924 -0.06121675
#> [55]  0.36269315  0.21511620 -0.10981132  0.70468008 -0.50725567 -0.45084839
#> [61] -0.89026271  0.99551738  0.57256009 -0.18205574 -0.78124191

isBetween(x, cos(x) + 1, cos(x) - 1)
#>   [1] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [13] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [25] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [37] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [49] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [61] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [73] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [85] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [97] FALSE FALSE FALSE FALSE