isBetween

Assess if x is between left and right bounds. Shortcut for x >= left & x <= right if strict=FALSE (default). Shortcut for x > left & x < right if strict=TRUE.

Usage

isBetween(x, left, right, strict = FALSE)

Arguments

x

Numeric values to assess

left

Numeric value(s) used as lower bound

right

Numeric value(s) used as upper bound

strict

Logical value defining if x is strictly between left and right. Default value is FALSE.

Value

Logical values

Examples

isBetween(1:12, 7, 9)
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE  TRUE FALSE FALSE FALSE

x <- rnorm(1e2)
x[isBetween(x, -1, 1)]
#>  [1] -0.152438067  0.298512751  0.063559267  0.406568360 -0.282245595
#>  [6] -0.523836734 -0.552400319 -0.068486888 -0.378293373  0.979934520
#> [11] -0.028690955  0.438797788 -0.304489409  0.643860304 -0.147879663
#> [16] -0.165361009 -0.224341646 -0.030511649 -0.832449756 -0.148391664
#> [21] -0.752472313 -0.420525084  0.890285093  0.373762054 -0.658804351
#> [26]  0.546994609 -0.033974541 -0.819479078  0.386583318 -0.321622643
#> [31] -0.500216846 -0.319204698 -0.929101964  0.002487332  0.177080079
#> [36] -0.206510105  0.727642025 -0.892573919  0.353756410 -0.919631658
#> [41]  0.558731230 -0.547128150  0.391178268 -0.448850804  0.245535788
#> [46] -0.633879555 -0.985775816  0.391636881  0.153324540  0.142145293
#> [51]  0.733478651 -0.749676923  0.573628082  0.052400095 -0.908521869
#> [56]  0.848789397  0.850324341 -0.170574422 -0.696995086 -0.146517175
#> [61]  0.128680917  0.191402161  0.254306966 -0.729321570  0.284944958
#> [66]  0.995456779 -0.530303273 -0.140217462 -0.679701116  0.996583633
#> [71] -0.269108533 -0.631655216  0.924668428 -0.797216503 -0.767073365
#> [76] -0.574026532

isBetween(x, cos(x) + 1, cos(x) - 1)
#>   [1] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [13] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [25] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [37] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [49] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [61] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [73] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [85] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [97] FALSE FALSE FALSE FALSE