isBetween

Assess if x is between left and right bounds. Shortcut for x >= left & x <= right if strict=FALSE (default). Shortcut for x > left & x < right if strict=TRUE.

Usage

isBetween(x, left, right, strict = FALSE)

Arguments

x

Numeric values to assess

left

Numeric value(s) used as lower bound

right

Numeric value(s) used as upper bound

strict

Logical value defining if x is strictly between left and right. Default value is FALSE.

Value

Logical values

Examples

isBetween(1:12, 7, 9)
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE  TRUE FALSE FALSE FALSE

x <- rnorm(1e2)
x[isBetween(x, -1, 1)]
#>  [1] -0.524223835 -0.455652121  0.158914583 -0.047586558 -0.965133369
#>  [6]  0.493774123 -0.036351903  0.442199038  0.105192910  0.498401591
#> [11] -0.076909066  0.222037906  0.577034284 -0.615185290 -0.447444578
#> [16] -0.698207158  0.725948528 -0.071739757  0.818059016 -0.578953904
#> [21] -0.517230339 -0.172830153  0.290502021 -0.089376971 -0.757683724
#> [26] -0.748253546  0.311342274  0.310517878 -0.068155380 -0.382166722
#> [31] -0.449262538 -0.330022742 -0.922272010 -0.881887725 -0.776430071
#> [36]  0.370418927  0.299656077 -0.716392211  0.649543319  0.347565026
#> [41] -0.432488445  0.121515285  0.674021667  0.212467907  0.292304528
#> [46] -0.049762670 -0.433364902  0.927692204  0.002083959 -0.286403872
#> [51]  0.343411043 -0.529217319 -0.280669076  0.850939087

isBetween(x, cos(x) + 1, cos(x) - 1)
#>   [1] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [13] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [25] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [37] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [49] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [61] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [73] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [85] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [97] FALSE FALSE FALSE FALSE