isBetween

Assess if x is between left and right bounds. Shortcut for x >= left & x <= right if strict=FALSE (default). Shortcut for x > left & x < right if strict=TRUE.

Usage

isBetween(x, left, right, strict = FALSE)

Arguments

x

Numeric values to assess

left

Numeric value(s) used as lower bound

right

Numeric value(s) used as upper bound

strict

Logical value defining if x is strictly between left and right. Default value is FALSE.

Value

Logical values

Examples

isBetween(1:12, 7, 9)
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE  TRUE FALSE FALSE FALSE

x <- rnorm(1e2)
x[isBetween(x, -1, 1)]
#>  [1]  0.462230890  0.721838525 -0.828908797 -0.271249446 -0.009989389
#>  [6] -0.807208543  0.686958591 -0.443275803 -0.779322851  0.985327849
#> [11] -0.864323795 -0.460939226 -0.655219590 -0.528268145 -0.049962639
#> [16]  0.904533685  0.132826968 -0.774901099 -0.514147177  0.542709705
#> [21] -0.756024387  0.761894428 -0.062578742  0.789465514 -0.389486354
#> [26] -0.171457364  0.101331129 -0.698152844  0.149247579 -0.226713545
#> [31] -0.902754092 -0.543245325  0.052568649 -0.390830374  0.612850592
#> [36] -0.452484714 -0.179720019 -0.049477823 -0.873438877 -0.548594215
#> [41] -0.154825780  0.466037872  0.192102108 -0.808115410 -0.934150305
#> [46]  0.472349011  0.817268152 -0.694176922 -0.451433386  0.898430957
#> [51] -0.997029187 -0.438122389 -0.747214078 -0.474629217  0.869426270
#> [56]  0.749260083  0.173519115 -0.988696160 -0.658606549 -0.230215629
#> [61]  0.095619316  0.147117584  0.183689649  0.879491900  0.406325837
#> [66] -0.391656885 -0.078064908  0.835284334 -0.475650250

isBetween(x, cos(x) + 1, cos(x) - 1)
#>   [1] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [13] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [25] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [37] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [49] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [61] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [73] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [85] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [97] FALSE FALSE FALSE FALSE