isBetween

Assess if x is between left and right bounds. Shortcut for x >= left & x <= right if strict=FALSE (default). Shortcut for x > left & x < right if strict=TRUE.

Usage

isBetween(x, left, right, strict = FALSE)

Arguments

x

Numeric values to assess

left

Numeric value(s) used as lower bound

right

Numeric value(s) used as upper bound

strict

Logical value defining if x is strictly between left and right. Default value is FALSE.

Value

Logical values

Examples

isBetween(1:12, 7, 9)
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE  TRUE FALSE FALSE FALSE

x <- rnorm(1e2)
x[isBetween(x, -1, 1)]
#>  [1] -0.19804296  0.21611240  0.22159558 -0.21925492  0.17305206  0.81462815
#>  [7] -0.33755593  0.34279953 -0.03105489  0.59983067  0.40685705  0.40199880
#> [13]  0.09414637  0.32552681 -0.24279071  0.47719083 -0.49503655  0.46913565
#> [19] -0.78754253  0.17660095 -0.83120684  0.80899650 -0.18047716 -0.05890551
#> [25] -0.65397265 -0.20338305  0.28063719  0.67042474 -0.86125533 -0.03864362
#> [31]  0.83273849 -0.93504848  0.53578073  0.09735461  0.03393285  0.22582139
#> [37]  0.56248696  0.46578343 -0.17666644  0.78833740 -0.06321155  0.40761582
#> [43]  0.99184851  0.73485063  0.52456821 -0.06869841 -0.07638104  0.08888158
#> [49]  0.06875735 -0.95430187 -0.49167222 -0.92401677  0.58844859 -0.21933807
#> [55] -0.86685403  0.23701544 -0.17989754 -0.10673334  0.52422127  0.23585688
#> [61] -0.68278895 -0.14422560 -0.27148728

isBetween(x, cos(x) + 1, cos(x) - 1)
#>   [1] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [13] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [25] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [37] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [49] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [61] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [73] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [85] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [97] FALSE FALSE FALSE FALSE