isBetween

Assess if x is between left and right bounds. Shortcut for x >= left & x <= right if strict=FALSE (default). Shortcut for x > left & x < right if strict=TRUE.

Usage

isBetween(x, left, right, strict = FALSE)

Arguments

x

Numeric values to assess

left

Numeric value(s) used as lower bound

right

Numeric value(s) used as upper bound

strict

Logical value defining if x is strictly between left and right. Default value is FALSE.

Value

Logical values

Examples

isBetween(1:12, 7, 9)
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE  TRUE FALSE FALSE FALSE

x <- rnorm(1e2)
x[isBetween(x, -1, 1)]
#>  [1] -0.357887047  0.232301225 -0.393209799  0.738194781  0.569890936
#>  [6]  0.084829768 -0.733871509  0.579905502 -0.956108790  0.591832255
#> [11]  0.813996181 -0.953844202 -0.685673697 -0.094761427 -0.419110324
#> [16]  0.402889418  0.759648475  0.453294791  0.401906226 -0.376727849
#> [21] -0.038550139 -0.041628998  0.505592396  0.618756617 -0.131227920
#> [26] -0.453029249  0.144818747 -0.288370247  0.761022984 -0.728162231
#> [31] -0.374967239 -0.105838055  0.670433240  0.525193194 -0.488171057
#> [36]  0.498198050 -0.605834725  0.686397691 -0.311241180  0.022285374
#> [41]  0.960964417  0.463002479  0.449558188  0.516587233  0.180619233
#> [46] -0.058658855  0.430946139 -0.194136204 -0.688612380 -0.173194765
#> [51] -0.035127328  0.632020023 -0.567536647 -0.913954964 -0.868992280
#> [56] -0.004540953  0.261247956  0.236879912  0.763452379 -0.612636721
#> [61]  0.594404805 -0.597453303 -0.020967303 -0.436466225 -0.247267733
#> [66] -0.609735165  0.845592863  0.583334093  0.572974877  0.385786092

isBetween(x, cos(x) + 1, cos(x) - 1)
#>   [1] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [13] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [25] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [37] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [49] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [61] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [73] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [85] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [97] FALSE FALSE FALSE FALSE