isBetween

Assess if x is between left and right bounds. Shortcut for x >= left & x <= right if strict=FALSE (default). Shortcut for x > left & x < right if strict=TRUE.

Usage

isBetween(x, left, right, strict = FALSE)

Arguments

x

Numeric values to assess

left

Numeric value(s) used as lower bound

right

Numeric value(s) used as upper bound

strict

Logical value defining if x is strictly between left and right. Default value is FALSE.

Value

Logical values

Examples

isBetween(1:12, 7, 9)
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE  TRUE FALSE FALSE FALSE

x <- rnorm(1e2)
x[isBetween(x, -1, 1)]
#>  [1] -0.067495700  0.897908235 -0.927529749 -0.014507618  0.315450144
#>  [6] -0.289659966 -0.322367626 -0.231259160 -0.005019611  0.371335171
#> [11]  0.844782053 -0.409495793  0.813169031 -0.657654998 -0.141054865
#> [16]  0.029496544 -0.368694069  0.736579114 -0.740272716 -0.352807607
#> [21]  0.027045690 -0.261567198  0.756818777  0.433903994  0.408906168
#> [26] -0.227999488 -0.230565912  0.439205877  0.955446903  0.459929958
#> [31] -0.917985872  0.025458305 -0.624004770  0.067172251 -0.910111920
#> [36]  0.799059913 -0.703898566  0.686888557  0.955727817 -0.084154465
#> [41] -0.247360747 -0.244306079 -0.891165807  0.266296165 -0.149821130
#> [46] -0.071256113 -0.867526506 -0.572938929 -0.862096219 -0.112877580
#> [51]  0.720307883  0.822572771  0.264676163 -0.451142609 -0.350301096
#> [56]  0.352672044 -0.325064490  0.850688651 -0.859119433 -0.219871013
#> [61]  0.412874703  0.165707510 -0.516694104 -0.242391477 -0.454692593
#> [66] -0.759120758  0.520511096 -0.219216366 -0.187650191  0.173375418
#> [71] -0.830440253 -0.730178223 -0.265265400  0.074569066 -0.040120748
#> [76]  0.220235156

isBetween(x, cos(x) + 1, cos(x) - 1)
#>   [1] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [13] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [25] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [37] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [49] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [61] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [73] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [85] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [97] FALSE FALSE FALSE FALSE