isBetween

Assess if x is between left and right bounds. Shortcut for x >= left & x <= right if strict=FALSE (default). Shortcut for x > left & x < right if strict=TRUE.

Usage

isBetween(x, left, right, strict = FALSE)

Arguments

x

Numeric values to assess

left

Numeric value(s) used as lower bound

right

Numeric value(s) used as upper bound

strict

Logical value defining if x is strictly between left and right. Default value is FALSE.

Value

Logical values

Examples

isBetween(1:12, 7, 9)
#>  [1] FALSE FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE  TRUE FALSE FALSE FALSE

x <- rnorm(1e2)
x[isBetween(x, -1, 1)]
#>  [1]  0.064324904 -0.647260180 -0.351311924  0.457896945 -0.645538798
#>  [6]  0.828208494  0.658149701 -0.036770797  0.507555782  0.851613757
#> [11]  0.356212732 -0.611804528 -0.155701843  0.528185155  0.439815109
#> [16] -0.724084586  0.334618254 -0.313341072 -0.496668499  0.652325834
#> [21] -0.078769172  0.125056328  0.860198033  0.217375383 -0.098318587
#> [26]  0.606959430 -0.042984435 -0.575964577  0.072906160  0.849512591
#> [31] -0.771948628 -0.963016337  0.212439125 -0.769049153 -0.402571695
#> [36]  0.521110770  0.337238417  0.441845801 -0.640852515  0.003522375
#> [41]  0.856663827  0.218641067 -0.758616001  0.604313530 -0.237881193
#> [46] -0.763735180  0.802877604 -0.253583709  0.384238751 -0.007280062
#> [51] -0.757323719 -0.151076399 -0.600396745 -0.141247018 -0.268289698
#> [56] -0.944665889 -0.625047271  0.534022441  0.469758098  0.249946088
#> [61]  0.878062273  0.516756237 -0.919478138  0.026488915  0.191016648
#> [66]  0.695092755 -0.116733104 -0.029225505 -0.073454379 -0.023628545
#> [71] -0.358767497  0.384148999 -0.493259735  0.379643167  0.681168275
#> [76]  0.870362107 -0.030318691 -0.938796496  0.498213088  0.686889438

isBetween(x, cos(x) + 1, cos(x) - 1)
#>   [1] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [13] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [25] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [37] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [49] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [61] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [73] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [85] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [97] FALSE FALSE FALSE FALSE